pyproject.toml
[project]
name = "bayz"
version = "0.1.0"
description = "Bayesian learning loop powered by LLMs observing a retro game simulation"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic-ai-slim[anthropic,logfire]>=0.2",
    "logfire[fastapi]>=3.0",
    "pygame-ce>=2.5",
    "numpy>=2.0",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "pydantic-evals>=0.1",
    "resend>=2.0",
    "itsdangerous>=2.1",
]
